animated pidgey line, added ability to merge in animations via xml, see pidgey line xmls for example.
the animation merging allows for easier making of animations, in the
case of the pidgey line, the folded wings are done by the idle
animation, and then the walking animation just deals with legs.  idle
animation also affects the legs, so the attribute merge="idle->walking"
makes it merge idle into walking, only adding to what walking did not
already animate.