added item storage AI
currently just stores in closest inventory to home position, it is an
IStorable as it will eventually keep track of a set of inventories to
store in.

Modified the hunger AI to also keep a stack of what it eats, doubling
the food from say eating a better.

Added initial template for a gather AI, which will gather items for
storage.