Added Pokecube Mobs
since users do not seem to be able to understand how to use vanilla
resource packs, I added a mod called Pokecube Mobs, which just wraps the
resource packs into a mod.

Also fixed the ability to register model providers from an external mod,
as well as adding support for models/textures from other mod locations.

the git.ignore ignores all of the assets, as they are just copies of the
ones in the gen specific folders.