Misc bug fixes and tweaks.
 -Modified the display info gui to allow the list to be above the name
tag, also made it not be able to go off screen.
 -Fixed the pokemobs spawning on wrong side of wall, at least for small
ones, if the pokemob is too large, it still may end up on wrong side
when released too close.
 -Fixed some of the cases where flying/fast pokemobs can phase through
walls, may result in them moving slowly near walls though, will need to
check.
 -Added v:x,y,z arguments to the /pokecube make command, it will
determine the relative location to the command block that the pokemob
spawns at.
 -Improved target detection for the attack command and for the throwing
of pokecubes.
 -Moved attacks to properly using location based targets, please let me
know if any attacks stop working due to this.
 -Made transform not-interceptable to make it work again after the above
change.