Stage 1 of the great config re-write.
Moving all of the configs over to one file, and added proper
localizations for them.

This makes them easier to manage via the ingame config gui, it also
facilitates in the addition of new config options for various things.

Next step is to make sure all of the options still work as intended.