Tweak trainer AI a bit, increase forget range. Tweaks to pokemob AI
Hopefully this will fix the issues with riding pokemobs falling into the
ground.
I also did some tweaks to contact move AI, it should be a bit more
responsive when used on other pokemobs.
Also some tweaks to movement when a pokemob is transformed.

Tweaks to threaded AI, let me know if any issues are noticed.