Fixes for graves mods.
Now it does the send to PC on the Drops event instead. This prevents
dupes from graves mods.
When a player picks up another player's pokecube from the ground, it
will be sent to owner's pc.  This fixes some other graves related
issues, It does not fix ones if the grave directly places in player's
inventory however.