Fixes issue with pokemobs going missing on death, some internal changes.
Refactored the system for adding in pokemobs from addons, now it allows
specificying the location of the models and textures per pokedex entry.

This allows Pokecube Mobs to use a more organized format for the models,
making it easier to keep track of what is in.

Refactored some of the onUpdate stuff in entities to be called from
AIRunnables instead. This will allow more cases where addons can
over-write the default AI.

Fixed issue where /pokemake results in the pokemob evolving immediately
after spawning.
Fixes the riddenoffsets not applying, improves on ridden offsets to be
more powerful, allowing offseting anywhere relative to the pokemob.
Fixes release issues in PC
Fixes a missing pokemob bug on death.
Adds sail fossil item