Re-organize some things with player data, add some debug commands.
Added /pokecube fixFromPC and /pokecube resetreward

'/pokecube fixFromPC <player>' will give achivements for pokemobs that
players have evolutions of in their PC, if they are missing the prevous
forme. This is to fix any players who had their acheivements lost due to
the update.

'/pokecube resetreward <player> <reward>' will reset the reward status
for that reward for the player.  This allows the player to try to get
the reward again.