Re-add trade with button to make mob wild, add trade with wild to own.
The trade with button for wild was removed for exploits related to
farming pokemobs, but that is now possible to do with eggs + hopper +
dispenser, so the mechanic for trading table was re-added.

This allows players to give another player a pokemob, by "releasing" it
by trading it for a button, then the other player can get ownership by
trading it for another button.