Massive additions to pokewatch gui.
Adds pokedex functionality to the pokewatch.
Includes pages for:
	Pokemob movesets.
	Allows swapping pokemobs moves via click and drag.
	Pokemob spawn locations.
	Pokemob stats.
	Base Stats for pokemobs not being inspected.
	Pokemob breeding groups.
	Mobs that spawn at current location.
	
Overhaul of the wiki pages section, now it properly culls the
whitespaces, making the books easier to read.  Also supports page goto
links in the books.
Added flag in book rewards to allow them to only show in the watch, and
not be given as an ingame item.
Supports clicking on the entries in the breeding groups to go to that
entries page.