Tweaks to trainer code for better customization and support.
Made it so that trainers can have various max numbers of pokemobs,
default trainers are still capped at 6.
Made relevant changes to trainer editor to account for this.
Fixed a bug with not being able to spawn trainers in certain cases.
Standardized the belt renderer for trainers, This makes it auto-add
renderer for any mob that is a trainer.  Also more properly arrange the
cubes based on max number of cubes the trainer can have.

Also re-arranged a debug print for item lookups