Major overhaul of AI state tracking.
AIStates have been split into the following:
GeneralStates
LogicStates
CombatStates

This split is made to better manage setting/getting of the states, as
well as to allow adding more additional states later, as the existing
system was hitting the limit of how much info can be crammed into an
int.

This will still be save compatible with old versions, as for a few
versions I will keep the converter from old states to new states, this
probably isn't too vital, as the worst that it should result in is
pokemobs going off stay/guard/sit.