prevents water mobs (like fish) from pathing out of water when idle.
This should result in things like karp sitting in their ponds, swimming
around there occasionally, but not randomly deciding to go for a flop
around out of the water.  They can still go flopping if they decide they
want to attack or eat something, just not due to idle movement.